EL via Patchmanger
Moves support for EL into using the new system by PatchManger for enabling/disabling patches.  As part of this, the EL patch has been split into four, each of which can be enabled/disabled independently.

 - MKS Classes can gain construction skill.
 - MKS Parts can have workshops/survey stations
 - The EL resource chain can be replaced with MKS’s resource chain
 - EL parts can be hidden from the game.

The default application of these patches is the same as before, so there should be no impact for users without PatchManger.